Make it better
Major changes:
 - Change input format (text appears after a begin time description) to
   make each line more independent (easier to process with other tools).
 - Change BEGIN/END format: parts can be omitted and the time and date
   are separate (so the beginning of today can be specified easily
   without knowing today's date, for example).
 - Do most date/time handling manually. This frees us from the
   limitation of not knowing the size of a time_t, and hence allows us
   to permit a wider range of dates, instead of conservatively bounding
   the year to [1970, 2037].
 - Add output format options to betters support years before 1AD.
 - Use suckless arg.h instead of stdlib getopt.
 - Improve internal error handling.
 - Cleanup iteration code (the 5 levels of nesting was a mess).
 - Improve code consistency and formatting.

DST handling and local timezone changes are slightly broken. For example,
the current code considers 2:00AM-2:59AM on 8 March 2020 to be valid
times with a America/Vancouver timezone, but, actually, those times never
existed, because that's when the switch from PST to PDT occurred. The
upside is that, unless the output format contains a %bu/%eu field, we
never have to consider DST/timezones, since we just do all time math with
the local time values, which simplifies things substantially.
(Currently, when the output format does contain a %bu/%eu field, we
defer to time.h, so it actually works properly.)